Extract Packeta label PDF helper
